Add check for latest stored version to deploy-docs.js

+++ theme/deploy-docs.js
@@ -1,4 +1,6 @@
 const util = require('util');
+const path = require('path');
+const fs = require('fs')
 const exec = util.promisify(require('child_process').exec);
 
 const version = require('../package.json').version;
@@ -22,14 +24,20 @@ async function start() {
   // Stash previous changes
   await run(`> .temp && git add . && git stash`, "Storing any changes to restore later")
   
-  // Generate stored doc files version
-  await run(`./node_modules/.bin/jsdoc -c jsdoc.json -d ./theme/static/versions/${version}/`, `Building doc files into: ./theme/static/versions/${version}/ directory...`)
+  const versions = await fs.readdirSync(path.join(__dirname, '/static/versions/'))
   
-  await run(`rm -rf theme/static/versions/${version}/cdn/`, "Cleaning up...")
-  
-  await run(`./node_modules/.bin/webpack`, "Generating latest CDN dist")
-  
-  await run(`git add . && git commit -am 'Update stored doc version ${version}'`, "Committing changes to git")
+  if(versions[Object.keys(versions).length-1] !== version) {
+    console.log('Missing latest package version from stored doc versions')
+    
+    // Generate stored doc files version
+    await run(`./node_modules/.bin/jsdoc -c jsdoc.json -d ./theme/static/versions/${version}/`, `Building doc files into: ./theme/static/versions/${version}/ directory...`)
+    
+    await run(`rm -rf theme/static/versions/${version}/cdn/`, "Cleaning up...")
+    
+    await run(`./node_modules/.bin/webpack`, "Generating latest CDN dist")
+    
+    await run(`git add . && git commit -am 'Update stored doc version ${version}'`, "Committing changes to git")
+  }
   
   // Generate current docfiles
   await run(`./node_modules/.bin/jsdoc -c jsdoc.json -d .`, 'Building latest doc files')
